+++ CHANGELOG.md
@@ -5,10 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project aims to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.1.1] - 2021-12-08
 ### Changed
 - `sampling.save_basis_transforms` now uses Pickle protocol 4 instead of 3 (which is the default).
 This enables the storage of transforms beyond the size of 4GB.
-
 ### Added
 - `sampling.draw_samples_from_mps` is implemented.